chore: extract builder and preview functions outside of dcl.ts
