+++ kernel/packages/unity-interface/setupPointerLock.ts
@@ -1,0 +1,8 @@
+import { unityInterface } from './dcl'
+export function setupPointerLock() {
+  document.addEventListener('pointerlockchange', e => {
+    if (!document.pointerLockElement) {
+      unityInterface.UnlockCursor()
+    }
+  })
+}
